Refactor action imports and enhance DeviceAgent configuration
- Updated action import paths from `powergrid.core.actions` to `powergrid.core.action`.
- Modified `DeviceAgent` to retrieve `action_callback` from `device_config`.
- Added `prefix_names` attribute and `_iter_ready_features` method to `DeviceState`.
- Adjusted `Grid`, `Transformer`, and `Storage` classes to pass `action_callback` in `device_config`.
- Added a new quick start script for environment setup quick test.
- Updated tests to reflect changes in action imports.
